Outsource functions to find conflicting products and their names
